Changed distList to List<Long>.
Seems like gremlin is stuffing longs in there... not Integers. This
becomes aparent when we try to access this list at line 290.

+++ src/main/java/org/ellitron/ldbc/driver/workloads/torc/db/TorcDb.java
@@ -201,7 +201,7 @@ public class TorcDb extends Db {
 
                 GraphTraversalSource g = graph.traversal();
 
-                List<Integer> distList = new ArrayList<>(resultLimit);
+                List<Long> distList = new ArrayList<>(resultLimit);
                 List<Vertex> matchList = new ArrayList<>(resultLimit);
 
                 Vertex root = g.V(new UInt128(Entity.PERSON.getNumber(), personId)).next();
